refactor: move FlowNodeInstanceMetadataBuilder out of rest
