+++ operate/webapp/src/main/java/io/camunda/operate/webapp/elasticsearch/reader/FlowNodeInstanceReader.java
@@ -53,7 +53,7 @@ import io.camunda.operate.store.ScrollException;
 import io.camunda.operate.util.ElasticsearchUtil;
 import io.camunda.operate.util.ElasticsearchUtil.QueryType;
 import io.camunda.operate.webapp.data.IncidentDataHolder;
-import io.camunda.operate.webapp.rest.FlowNodeInstanceMetadataBuilder;
+import io.camunda.operate.webapp.reader.metadata.FlowNodeInstanceMetadataBuilder;
 import io.camunda.operate.webapp.rest.dto.activity.FlowNodeInstanceDto;
 import io.camunda.operate.webapp.rest.dto.activity.FlowNodeInstanceQueryDto;
 import io.camunda.operate.webapp.rest.dto.activity.FlowNodeInstanceRequestDto;
