+++ operate/webapp/src/main/java/io/camunda/operate/webapp/opensearch/reader/OpensearchFlowNodeInstanceReader.java
@@ -45,7 +45,7 @@ import io.camunda.operate.util.CollectionUtil;
 import io.camunda.operate.webapp.data.IncidentDataHolder;
 import io.camunda.operate.webapp.elasticsearch.reader.ProcessInstanceReader;
 import io.camunda.operate.webapp.reader.FlowNodeInstanceReader;
-import io.camunda.operate.webapp.rest.FlowNodeInstanceMetadataBuilder;
+import io.camunda.operate.webapp.reader.metadata.FlowNodeInstanceMetadataBuilder;
 import io.camunda.operate.webapp.rest.dto.activity.FlowNodeInstanceDto;
 import io.camunda.operate.webapp.rest.dto.activity.FlowNodeInstanceQueryDto;
 import io.camunda.operate.webapp.rest.dto.activity.FlowNodeInstanceRequestDto;
