+++ operate/webapp/src/test/java/io/camunda/operate/webapp/reader/metadata/FlowNodeInstanceMetadataBuilderTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.operate.webapp.rest;
+package io.camunda.operate.webapp.reader.metadata;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.mockito.Mockito.when;
